docs: v0.25.0 release prep — naga v0.17.4, DXIL docs, lint fixes
- CHANGELOG: add GLES Y-flip, DXIL sampler heap, pipeline error logging
- README: add Environment Variables section (GOGPU_DX12_DXIL)
- ROADMAP: update current state to v0.25.0, refresh milestones
- go.mod: naga v0.17.3 → v0.17.4 (DXIL full rendering)
- Fix gosec G306 (file permissions 0o644 → 0o600)
- Fix gocritic exitAfterDefer (os.Exit → return error)
- Fix nestif in pipeline.go (extract resolveShaderBytecode helper)
- Fix gofmt in device.go (inline closure formatting)

+++ hal/gles/device_linux.go
@@ -130,7 +130,7 @@ func (d *Device) UnmapBuffer(buffer hal.Buffer) error {
 	}
 	if buf.usage&gputypes.BufferUsageMapWrite != 0 && buf.id != 0 {
 		d.glCtx.BindBuffer(buf.target, buf.id)
-		d.glCtx.BufferSubData(buf.target, 0, len(buf.mapped), unsafe.Pointer(&buf.mapped[0]))
+		d.glCtx.BufferSubData(buf.target, 0, len(buf.mapped), uintptr(unsafe.Pointer(&buf.mapped[0])))
 		d.glCtx.BindBuffer(buf.target, 0)
 	}
 	buf.mapped = nil
